+++ src/loader/filetypes/PluginFile.js
@@ -10,7 +10,6 @@ var File = require('../File');
 var FileTypesManager = require('../FileTypesManager');
 var GetFastValue = require('../../utils/object/GetFastValue');
 var IsPlainObject = require('../../utils/object/IsPlainObject');
-var PluginManager = require('../../plugins/PluginManager');
 
 /**
  * @typedef {object} Phaser.Loader.FileTypes.PluginFileConfig